adjusted code to agreed code-style
