+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ExploreV2/Scripts/Sections/PlacesAndEventsSection/SubSections/HighlightsSubSection/HighlightsSubSectionMenu/HighlightsSubSectionComponentView.cs
@@ -1,149 +1,21 @@
 using DCL;
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using Cysharp.Threading.Tasks;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
 
-public interface IHighlightsSubSectionComponentView
-{
-    /// <summary>
-    /// It will be triggered when all the UI components have been fully initialized.
-    /// </summary>
-    event Action OnReady;
-
-    /// <summary>
-    /// It will be triggered when the place info button is clicked.
-    /// </summary>
-    event Action<PlaceCardComponentModel> OnPlaceInfoClicked;
-
-    /// <summary>
-    /// It will be triggered when the event info button is clicked.
-    /// </summary>
-    event Action<EventCardComponentModel> OnEventInfoClicked;
-
-    /// <summary>
-    /// It will be triggered when the place JumpIn button is clicked.
-    /// </summary>
-    event Action<HotScenesController.HotSceneInfo> OnPlaceJumpInClicked;
-
-    /// <summary>
-    /// It will be triggered when the event JumpIn button is clicked.
-    /// </summary>
-    event Action<EventFromAPIModel> OnEventJumpInClicked;
-
-    /// <summary>
-    /// It will be triggered when the subscribe event button is clicked.
-    /// </summary>
-    event Action<string> OnEventSubscribeEventClicked;
-
-    /// <summary>
-    /// It will be triggered when the unsubscribe event button is clicked.
-    /// </summary>
-    event Action<string> OnEventUnsubscribeEventClicked;
-
-    /// <summary>
-    /// It will be triggered when the view all events button is clicked.
-    /// </summary>
-    event Action OnViewAllEventsClicked;
-
-    /// <summary>
-    /// It will be triggered when a new friend handler is added by a place card.
-    /// </summary>
-    event Action<FriendsHandler> OnFriendHandlerAdded;
-
-    /// <summary>
-    /// It will be triggered each time the view is enabled.
-    /// </summary>
-    event Action OnHighlightsSubSectionEnable;
-
-    /// <summary>
-    /// Colors used for the background of the friends heads.
-    /// </summary>
-    Color[] currentFriendColors { get; }
-
-    /// <summary>
-    /// Set the trending places/events component with a list of places and events.
-    /// </summary>
-    /// <param name="places">List of places (model) to be loaded.</param>
-    /// <param name="events">List of events (model) to be loaded.</param>
-    void SetTrendingPlacesAndEvents(List<PlaceCardComponentModel> places, List<EventCardComponentModel> events);
-
-    /// <summary>
-    /// Set the trending places and events component in loading mode.
-    /// </summary>
-    /// <param name="isVisible">True for activating the loading mode.</param>
-    void SetTrendingPlacesAndEventsAsLoading(bool isVisible);
-
-    /// <summary>
-    /// Set the featured places component with a list of places.
-    /// </summary>
-    /// <param name="places">List of places (model) to be loaded.</param>
-    void SetFeaturedPlaces(List<PlaceCardComponentModel> places);
-
-    /// <summary>
-    /// Set the featured places component in loading mode.
-    /// </summary>
-    /// <param name="isVisible">True for activating the loading mode.</param>
-    void SetFeaturedPlacesAsLoading(bool isVisible);
-
-    /// <summary>
-    /// Set the live events component with a list of places.
-    /// </summary>
-    /// <param name="events">List of events (model) to be loaded.</param>
-    void SetLiveEvents(List<EventCardComponentModel> events);
-
-    /// <summary>
-    /// Set the live events component in loading mode.
-    /// </summary>
-    /// <param name="isVisible">True for activating the loading mode.</param>
-    void SetLiveAsLoading(bool isVisible);
-
-    /// <summary>
-    /// Shows the Place Card modal with the provided information.
-    /// </summary>
-    /// <param name="placeInfo">Place (model) to be loaded in the card.</param>
-    void ShowPlaceModal(PlaceCardComponentModel placeInfo);
-
-    /// <summary>
-    /// Hides the Place Card modal.
-    /// </summary>
-    void HidePlaceModal();
-
-    /// <summary>
-    /// Shows the Event Card modal with the provided information.
-    /// </summary>
-    /// <param name="eventInfo">Event (model) to be loaded in the card.</param>
-    void ShowEventModal(EventCardComponentModel eventInfo);
-
-    /// <summary>
-    /// Hides the Event Card modal.
-    /// </summary>
-    void HideEventModal();
-
-    /// <summary>
-    /// Set the current scroll view position to 1.
-    /// </summary>
-    void RestartScrollViewPosition();
-
-    /// <summary>
-    /// Configure the needed pools for the places and events instantiation.
-    /// </summary>
-    void ConfigurePools();
-}
-
 public class HighlightsSubSectionComponentView : BaseComponentView, IHighlightsSubSectionComponentView
 {
-    internal const string TRENDING_PLACE_CARDS_POOL_NAME = "Highlights_TrendingPlaceCardsPool";
-    internal const int TRENDING_PLACE_CARDS_POOL_PREWARM = 10;
-    internal const string TRENDING_EVENT_CARDS_POOL_NAME = "Highlights_TrendingEventCardsPool";
-    internal const int TRENDING_EVENT_CARDS_POOL_PREWARM = 10;
-    internal const string FEATURED_PLACE_CARDS_POOL_NAME = "Highlights_FeaturedPlaceCardsPool";
-    internal const int FEATURED_PLACE_CARDS_POOL_PREWARM = 9;
-    internal const string LIVE_EVENT_CARDS_POOL_NAME = "Highlights_LiveEventCardsPool";
-    internal const int LIVE_EVENT_CARDS_POOL_PREWARM = 3;
+    private const string TRENDING_PLACE_CARDS_POOL_NAME = "Highlights_TrendingPlaceCardsPool";
+    private const int TRENDING_PLACE_CARDS_POOL_PREWARM = 10;
+    private const string TRENDING_EVENT_CARDS_POOL_NAME = "Highlights_TrendingEventCardsPool";
+    private const int TRENDING_EVENT_CARDS_POOL_PREWARM = 10;
+    private const string FEATURED_PLACE_CARDS_POOL_NAME = "Highlights_FeaturedPlaceCardsPool";
+    private const int FEATURED_PLACE_CARDS_POOL_PREWARM = 9;
+    private const string LIVE_EVENT_CARDS_POOL_NAME = "Highlights_LiveEventCardsPool";
+    private const int LIVE_EVENT_CARDS_POOL_PREWARM = 3;
 
     private readonly Queue<Func<UniTask>> cardsVisualUpdateBuffer = new Queue<Func<UniTask>>();
     private readonly Queue<Func<UniTask>> poolsIterativePrewarBuffer = new Queue<Func<UniTask>>();
@@ -170,6 +42,17 @@ public class HighlightsSubSectionComponentView : BaseComponentView, IHighlightsS
     [SerializeField] internal Color[] friendColors = null;
 
     [SerializeField] private Canvas canvas;
+    
+    internal PlaceCardComponentView placeModal;
+    internal EventCardComponentView eventModal;
+    internal Pool trendingPlaceCardsPool;
+    internal Pool trendingEventCardsPool;
+    internal Pool featuredPlaceCardsPool;
+    internal Pool liveEventCardsPool;
+    
+    private bool isUpdatingCardsVisual;
+    
+    public Color[] currentFriendColors => friendColors;
 
     public event Action OnReady;
     public event Action<PlaceCardComponentModel> OnPlaceInfoClicked;
@@ -181,17 +64,6 @@ public class HighlightsSubSectionComponentView : BaseComponentView, IHighlightsS
     public event Action OnViewAllEventsClicked;
     public event Action<FriendsHandler> OnFriendHandlerAdded;
     public event Action OnHighlightsSubSectionEnable;
-
-    internal PlaceCardComponentView placeModal;
-    internal EventCardComponentView eventModal;
-    internal Pool trendingPlaceCardsPool;
-    internal Pool trendingEventCardsPool;
-    internal Pool featuredPlaceCardsPool;
-    internal Pool liveEventCardsPool;
-
-    public Color[] currentFriendColors => friendColors;
-
-    private bool isUpdatingCardsVisual;
 
     public override void Start()
     {
@@ -207,7 +79,10 @@ public class HighlightsSubSectionComponentView : BaseComponentView, IHighlightsS
         OnReady?.Invoke();
     }
 
-    public override void OnEnable() { OnHighlightsSubSectionEnable?.Invoke(); }
+    public override void OnEnable()
+    {
+        OnHighlightsSubSectionEnable?.Invoke();
+    }
 
     public void SetActive(bool isActive)
     {
@@ -257,24 +132,54 @@ public class HighlightsSubSectionComponentView : BaseComponentView, IHighlightsS
         liveEvents.RefreshControl();
     }
 
-    private void UpdateCardsVisual()
+    public void SetTrendingPlacesAndEventsAsLoading(bool isVisible)
     {
-        if (!isUpdatingCardsVisual)
-            UpdateCardsVisualProcess().Forget();
+        SetTrendingPlacesAndEventsActive(!isVisible);
+        trendingPlacesAndEventsLoading.SetActive(isVisible);
     }
 
-    private async UniTask UpdateCardsVisualProcess()
+    internal void SetTrendingPlacesAndEventsActive(bool isActive) =>
+        trendingPlacesAndEvents.gameObject.SetActive(isActive);
+
+    public void SetFeaturedPlacesAsLoading(bool isVisible)
     {
-        isUpdatingCardsVisual = true;
+        featuredPlaces.gameObject.SetActive(!isVisible);
+        featuredPlacesLoading.SetActive(isVisible);
 
-        while (cardsVisualUpdateBuffer.Count > 0)
-            await cardsVisualUpdateBuffer.Dequeue().Invoke();
-
-        while (poolsIterativePrewarBuffer.Count > 0)
-            await poolsIterativePrewarBuffer.Dequeue().Invoke();
-
-        isUpdatingCardsVisual = false;
+        if (isVisible)
+            featuredPlacesNoDataText.gameObject.SetActive(false);
     }
+
+    public void SetLiveAsLoading(bool isVisible)
+    {
+        liveEvents.gameObject.SetActive(!isVisible);
+        liveEventsLoading.SetActive(isVisible);
+        viewAllEventsButton.gameObject.SetActive(!isVisible);
+
+        if (isVisible)
+            liveEventsNoDataText.gameObject.SetActive(false);
+    }
+
+    public void ShowPlaceModal(PlaceCardComponentModel placeInfo)
+    {
+        placeModal.Show();
+        ExplorePlacesUtils.ConfigurePlaceCard(placeModal, placeInfo, OnPlaceInfoClicked, OnPlaceJumpInClicked);
+    }
+
+    public void HidePlaceModal() => 
+        placeModal.Hide();
+
+    public void ShowEventModal(EventCardComponentModel eventInfo)
+    {
+        eventModal.Show();
+        ExploreEventsUtils.ConfigureEventCard(eventModal, eventInfo, OnEventInfoClicked, OnEventJumpInClicked, OnEventSubscribeEventClicked, OnEventUnsubscribeEventClicked);
+    }
+
+    public void HideEventModal() => 
+        eventModal.Hide();
+
+    public void RestartScrollViewPosition() => 
+        scrollView.verticalNormalizedPosition = 1;
 
     public void SetFeaturedPlaces(List<PlaceCardComponentModel> places)
     {
@@ -377,50 +282,23 @@ public class HighlightsSubSectionComponentView : BaseComponentView, IHighlightsS
 
         poolsIterativePrewarBuffer.Enqueue(() => trendingPlaceCardsPool.IterativePrewarm(HighlightsSubSectionComponentController.DEFAULT_NUMBER_OF_TRENDING_PLACES));
     }
-
-    public void SetTrendingPlacesAndEventsAsLoading(bool isVisible)
+    
+    private void UpdateCardsVisual()
     {
-        SetTrendingPlacesAndEventsActive(!isVisible);
-        trendingPlacesAndEventsLoading.SetActive(isVisible);
+        if (!isUpdatingCardsVisual)
+            UpdateCardsVisualProcess().Forget();
+        
+        async UniTask UpdateCardsVisualProcess()
+        {
+            isUpdatingCardsVisual = true;
+
+            while (cardsVisualUpdateBuffer.Count > 0)
+                await cardsVisualUpdateBuffer.Dequeue().Invoke();
+
+            while (poolsIterativePrewarBuffer.Count > 0)
+                await poolsIterativePrewarBuffer.Dequeue().Invoke();
+
+            isUpdatingCardsVisual = false;
+        }
     }
-
-    internal void SetTrendingPlacesAndEventsActive(bool isActive) =>
-        trendingPlacesAndEvents.gameObject.SetActive(isActive);
-
-    public void SetFeaturedPlacesAsLoading(bool isVisible)
-    {
-        featuredPlaces.gameObject.SetActive(!isVisible);
-        featuredPlacesLoading.SetActive(isVisible);
-
-        if (isVisible)
-            featuredPlacesNoDataText.gameObject.SetActive(false);
-    }
-
-    public void SetLiveAsLoading(bool isVisible)
-    {
-        liveEvents.gameObject.SetActive(!isVisible);
-        liveEventsLoading.SetActive(isVisible);
-        viewAllEventsButton.gameObject.SetActive(!isVisible);
-
-        if (isVisible)
-            liveEventsNoDataText.gameObject.SetActive(false);
-    }
-
-    public void ShowPlaceModal(PlaceCardComponentModel placeInfo)
-    {
-        placeModal.Show();
-        ExplorePlacesUtils.ConfigurePlaceCard(placeModal, placeInfo, OnPlaceInfoClicked, OnPlaceJumpInClicked);
-    }
-
-    public void HidePlaceModal() { placeModal.Hide(); }
-
-    public void ShowEventModal(EventCardComponentModel eventInfo)
-    {
-        eventModal.Show();
-        ExploreEventsUtils.ConfigureEventCard(eventModal, eventInfo, OnEventInfoClicked, OnEventJumpInClicked, OnEventSubscribeEventClicked, OnEventUnsubscribeEventClicked);
-    }
-
-    public void HideEventModal() { eventModal.Hide(); }
-
-    public void RestartScrollViewPosition() { scrollView.verticalNormalizedPosition = 1; }
 }
